modularize gather
